requirements.txt
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
anyio==3.7.1
asgiref==3.7.2
attrs==23.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
datasets==2.15.0
dill==0.3.7
fastapi==0.73.0
fastcore==1.5.29
filelock==3.13.1
fonttools==4.46.0
frozenlist==1.4.1
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.40
h11==0.14.0
huggingface-hub==0.19.4
idna==3.6
importlib-metadata==6.11.0
iniconfig==2.0.0
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
networkx==3.2.1
numpy==1.26.2
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
pluggy==1.3.0
protobuf==4.25.1
pyarrow==14.0.1
pyarrow-hotfix==0.6
pydantic==1.10.13
pydeck==0.8.1b0
Pygments==2.17.2
pyparsing==3.1.1
pytest==7.4.3
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
requests==2.31.0
rich==13.7.0
rpds-py==0.15.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
starlette==0.17.1
streamlit==1.29.0
sympy==1.12
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
torch==2.1.2
torchvision==0.16.2
tornado==6.4
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
uvicorn==0.15.0
validators==0.22.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0
